==> 2.cd138_spectra.expression_risk.Rmd <==
---
title: "CD138+ spectra and established expression-based risk scores"
author: "Rosalie Griffin Waller"
date: "27-OCT-2020"
output: pdf_document
---

```{r load_packages, message=F, include=T,error=F}
# Install and load required R packages
library(dplyr)
library(data.table)
library(ggplot2)
library("survival")
library("survminer")
```

Define data directory
```{r}
data_dir = "/path/to/data" # exclude ending "/"
```
```{r, include=FALSE}
# real data directory
data_dir = "/Users/rosal/OneDrive - University of Utah/2020/analyze/data/spectra.v2/data"
```

Load data
```{r}
# read in normalized and combat adjusted expression data on baseline samples
exp_cbat = read.csv(file = paste0(data_dir,
                                  "/baseline-expression-norm-combat.csv"))

# read in spectra and clinical data, remove clinical data
spectra = read.csv(file = paste0(data_dir,
                                  "/baseline-clinical-spectra-sd.csv")
                    ) %>% dplyr::select("SEQ_ID",starts_with("PC"))

dt.all = read.csv(paste0(data_dir,"/baseline-clinical-spectra-sd.csv"))
```

### 1. University of Arkansas UAMS-70 gene panel to classify patients as low or high risk for relapse
DOI: 10.1182/blood-2006-07-038430

1.1. **COMPUTE SCORE IN COMMPASS DATA**

Genes not found in data

Down-regulated: "PNPLA4","KIAA1754","MCLC","AD-020","PARG1","RFP2","FLJ20489";

Up-regulated: "FABP5","PDHA1","TRIP13","SELI","SLCI19A1","ASPM","STK6",
"FLJ13052","LAS1L","BIRC5","CKAP1","MGC57827","DKFZp779O175","PAPD1","EIF2C2",
"MGC4308","DSG2","C6orf173","MGC15606","KIF14","DKFZP586L0724","WEE1","ROBO1",
"MPHOSPH1"

```{r}
# working from ComBat adjusted expression estimates in baseline samples
DAT = exp_cbat 

# LIST OF UP-REGULATED GENES IN UAMS-70 GENE SCORE
up = c("FABP5","PDHA1","TRIP13","AIM2","SELI","SLCI19A1","LARS2",
       "OPN3","ASPM","CCT2","UBE2I","STK6","FLJ13052",
        "LAS1L","BIRC5","RFC4","CKS1B","CKAP1","MGC57827",
       "DKFZp779O175","PFN1","ILF3","IFI16","TBRG4","PAPD1",
        "EIF2C2","MGC4308","ENO1","DSG2","C6orf173","EXOSC4",
       "TAGLN2","RUVBL1","ALDOA","CPSF3","MGC15606","LGALS1",
        "RAD18","SNX5","PSMD4","RAN","KIF14","CBX3","TMPO",
       "DKFZP586L0724","WEE1","ROBO1","TCOF1","YWHAZ",
        "MPHOSPH1")
print(paste0("UAMS up regulated genes: ",
             length(intersect(up,colnames(DAT)))," of ",
             length(up)," genes in dataset"))
# select up regulated genes in data
anno_up = DAT %>% dplyr::select(intersect(up,colnames(DAT))) 

# LIST OF DOWN-REGULATED GENES IN UAMS-70 GENE SCORE
down = c("GNG10","PNPLA4","KIAA1754","AHCYL1","MCLC","EVI5","AD-020",
         "PARG1","CTBS","UBE2R2","FUCA1","RFP2","FLJ20489","LTBP1","TRIM33")
print(paste0("UAMS down regulated genes: ",
             length(intersect(down,colnames(DAT))),
             " of ",length(down)," genes in dataset"))
# select up regulated genes in data
anno_dw = DAT %>% dplyr::select(intersect(down,colnames(DAT))) 

# COMPUTE GEOMETRIC MEANS
x = DAT[,"SEQ_ID"] %>% data.table
colnames(x) = "SEQ_ID"
x$up = anno_up %>% rowMeans()
x$dw = anno_dw %>% rowMeans()

# COMPUTE PROPORTION OF MEAN UP/DOWN AND PLOT
x$score = x$up - x$dw # Note: expression already in log2 scale
#hist(x$score,breaks = 200,main = "Risk Score",
#xlab = "log2(mean up reg) - log2(mean down reg)")

#mean(x$score)
rm(DAT,anno_up,up,down,anno_dw) # Cleanup variables
```
**Clustering**
```{r}
# K-Means Cluster Analysis
fit <- kmeans(x$score, 3)

# get cluster max
aggregate(x$score,by=list(fit$cluster),FUN=max)

# append cluster assignment
x.score <- data.frame(x, fit$cluster)

# proportion in high/low risk group
#nrow(x.score[x.score$score > 2.424285,])/nrow(x.score)
#nrow(x.score[x.score$score < 2.424285,])/nrow(x.score)
```

**Histogram**
```{r}
h <- hist(x.score$score,breaks = 150,main = "",
          xlab = expression("Log"[2]*" Geometric Mean Ratio"))#,ylim = c(0,18))
abline(v=2.424285, col="purple",lwd=3)
text(x=1,y=18,labels = "80.4%")
text(x=3,y=18,labels = "19.6%")
text(x=2.42,y=20,labels = "2.42")
```


1.2. **LINEAR REGRESSION WITH SPECTRA**
```{r}
dt.70 = merge(x,spectra,by="SEQ_ID") # Merge PCs with computed UAMS risk score
# run linear regression with UAMS risk score as dependent variable
lm.70 = lm(data = dt.70[,-c("SEQ_ID","up","dw")], formula = score ~ .) 
summary(lm.70)

nsig = data.table(summary(lm.70)$coeff[-1,"Pr(>|t|)"]) %>% 
  subset(V1<0.05) %>% nrow() # Count sig spectra in model
print(paste0(nsig," of the ",ncol(dt.70[,-c("SEQ_ID","up","dw")])-1,
             " spectra were significat at p < 0.05 in the combined model"))

ajr2 = round(summary(lm.70)$adj.r.squared,digits = 2) # Print adjusted R^2
df1 = round(summary(lm.70)$fstatistic[2])
df2 = round(summary(lm.70)$fstatistic[3])
f = round(summary(lm.70)$fstatistic[1],digits = 1)
# Compute p-value from f-statistic
p = formatC(pf(f,df1,df2,lower.tail = F), format = "E", digits = 1) 
print(paste0("adjusted R^2=",ajr2,", F(",df1,",",df2,")=",f,", p=",p))

# Overall p-value
df1 = summary(lm.70)$fstatistic[2]
df2 = summary(lm.70)$fstatistic[3]
f = summary(lm.70)$fstatistic[1]
lm.70$p = pf(f,df1,df2,lower.tail = F) # Compute p-value from f-statistic
```

1.3. **Logistic regression, high v low risk score**
```{r}
dt.70$bin = as.factor(if_else(dt.70$score<2.424285,"low","high"))
# run logistic regression with UAMS risk score as dependent variable
glm.70 = glm(data = dt.70[,-c("SEQ_ID","up","dw","score")], 
             formula = bin ~ ., family = "binomial") 
summary(glm.70)

nsig = data.table(summary(glm.70)$coeff[-1,"Pr(>|z|)"]) %>% 
  subset(V1<0.05) %>% nrow() # Count sig spectra in model
print(paste0(nsig," of the ",ncol(dt.70[,-c("SEQ_ID","up","dw","score")])-1,
             " spectra were significat at p < 0.05 in the combined model"))

# Overall p-value
NLL = glm(data = dt.70[,-c("SEQ_ID","up","dw","score")], formula = bin ~ 1, family = "binomial")

pchisq(deviance(NLL)-deviance(glm.70),
                         df.residual(NLL)-df.residual(glm.70),
                         lower.tail=FALSE)
```

1.3.1 **High v low Survival Kaplan-Meier curves**
```{r}
pdt = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttcos","censos")])
pdt$yrs = pdt$ttcos/365

fit = survfit(Surv(time = pdt$yrs, event = pdt$censos) ~ bin, data=pdt)
temp.surv = summary(fit,times=c(1,3,5))
data.table(time=temp.surv$time,strata=temp.surv$strata,survival=temp.surv$surv)

p1 = ggsurvplot(survfit(Surv(time = pdt$yrs, event = pdt$censos) ~ bin, data=pdt),
           conf.int = TRUE,
           break.x.by = 1, xlab = "years",
           ggtheme = theme_set(theme_classic() + 
                                 theme(axis.title=element_text(size=11),
                                       axis.text=element_text(size=9),
                                       panel.grid.major.x = element_line(
                                         colour = "gray",linetype = 3)
                                 )
                               )
           )
p1

survdiff(formula = Surv(time = pdt$ttcos, event = pdt$censos) ~ bin, data=pdt)
```
UAMS High/Low and TTF
```{r}
pdt = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttctf1","censtf1")])
pdt$yrs = pdt$ttctf1/365

fit = survfit(Surv(time = pdt$yrs, event = pdt$censtf1) ~ bin, data=pdt)
temp.surv = summary(fit,times=c(1,3,5))
data.table(time=temp.surv$time,strata=temp.surv$strata,survival=temp.surv$surv)

p1 = ggsurvplot(survfit(Surv(time = pdt$yrs, event = pdt$censtf1) ~ bin, data=pdt),
           conf.int = TRUE,
           break.x.by = 1, xlab = "years",
           ggtheme = theme_set(theme_classic() + 
                                 theme(axis.title=element_text(size=11),
                                       axis.text=element_text(size=9),
                                       panel.grid.major.x = element_line(
                                         colour = "gray",linetype = 3)
                                 )
                               )
           )
p1

survdiff(formula = Surv(time = pdt$ttctf1, event = pdt$censtf1) ~ bin, data=pdt)
```

**OS at 1 year**
```{r}
dat = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttcos","censos")])
dat$category = if_else(dat$ttcos<=365 & dat$censos==1,"dead",
                       if_else(dat$ttcos>365,"alive","censored")) 

# number of events
evts = dat %>% filter(censos==1) %>% group_by(bin) %>% 
  count(censos,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
#cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("UAMS high/low OS HR at 1 year = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```
**OS at 3 years**
```{r}
dat = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttcos","censos")])
dat$category = if_else(dat$ttcos<=365*3 & dat$censos==1,"dead",
                       if_else(dat$ttcos>365*3,"alive","censored")) 

# number of events
evts = dat %>% filter(censos==1) %>% group_by(bin) %>% 
  count(censos,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
#cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("UAMS high/low OS HR at 3 years = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```
**OS at 5 years**
```{r}
dat = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttcos","censos")])
dat$category = if_else(dat$ttcos<=365*5 & dat$censos==1,"dead",
                       if_else(dat$ttcos>365*5,"alive","censored")) 

# number of events
evts = dat %>% filter(censos==1) %>% group_by(bin) %>% 
  count(censos,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
#cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("UAMS high/low OS HR at 5 years = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```


**TTF at 1 year**
```{r}
dat = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttctf1","censtf1")])
dat$category = if_else(dat$ttctf1<=365 & dat$censtf1==1,"dead",
                       if_else(dat$ttctf1>365,"alive","censored")) 

# number of events
evts = dat %>% filter(censtf1==1) %>% group_by(bin) %>% 
  count(censtf1,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
#cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("UAMS high/low TTF HR at 1 year = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```

**TTF at 3 years**
```{r}
dat = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttctf1","censtf1")])
dat$category = if_else(dat$ttctf1<=365*3 & dat$censtf1==1,"dead",
                       if_else(dat$ttctf1>365*3,"alive","censored")) 

# number of events
evts = dat %>% filter(censtf1==1) %>% group_by(bin) %>% 
  count(censtf1,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("UMAS high/low TTF HR at 3 years = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```


**TTF at 5 years**
```{r}
dat = merge(dt.70[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttctf1","censtf1")])
dat$category = if_else(dat$ttctf1<=365*5 & dat$censtf1==1,"dead",
                       if_else(dat$ttctf1>365*5,"alive","censored")) 

# number of events
evts = dat %>% filter(censtf1==1) %>% group_by(bin) %>% 
  count(censtf1,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("UMAS high/low TTF HR at 5 years = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```

1.4. **PLOT**

```{r}
theme_set(theme_classic() + 
            theme(legend.position="none",
                  legend.title = element_text(size=8),
                  legend.text = element_text(size=7),
                  axis.title=element_text(size=9),
                  axis.text=element_text(size=8)))
```

Actual v predicted values
```{r}
fit = lm.70
actual_preds <- data.table(cbind(dt.70[,c("SEQ_ID","bin")],
                                 actual=fit$model$score,predicted=fit$fitted.values))
cor(actual_preds[,-c("SEQ_ID","bin")])
# mean absolute percentage error
mean(abs((actual_preds$predicted - actual_preds$actual))/actual_preds$actual) 
```
Plot fitted values x actual values
```{r}
p1 = ggplot(actual_preds,aes(y = predicted, x = actual)) +
  geom_point(color = "#0033CC60") +
  xlab("UAMS-70 Score") + ylab("UAMS-70 PSL Score")
p1
ggsave(filename = "plots/UAMS-70_dot.pdf",plot = p1,width=3.75,height=3.75)
```

Waterfall High/Low risk score
```{r}
setorder(actual_preds,-predicted)
# predicted score - high risk cutoff from clustering
w <- ggplot(actual_preds,aes(y=predicted-2.424285,x=1:nrow(actual_preds),fill=bin)) + 
  geom_bar(stat="identity") +
  scale_fill_manual(values = c("blue","gray"), 
                    name = "UAMS-70 Risk Group", 
                    labels = c("High (n=151)", "Low (n=616)")) +
  xlab("Patients (N=767)") + ylab("UAMS-70 PSL Score") +
  scale_x_continuous(limits = c(0,nrow(actual_preds)+1), 
                     breaks=seq(1,nrow(actual_preds),1)) +
  theme(legend.position = c(1,1), legend.justification = c(1,1),
        axis.text = element_blank(), axis.ticks = element_blank())
w
ggsave(filename = "plots/UAMS-70_waterfall.pdf",plot = w,width=3.75,height=2)
```


### 2. Br J Haematol. 2020 May 15. Development of a RNA sequencing-based prognostic gene signature in multiple myeloma
DOI: 10.1111/bjh.16744

Developed at the Shahid Bahonar University of Kerman, Kerman, Iran. 
I will refer to this prognostic score as SBUK-17.

Genes and regression coefficients from Table III.

2.1. **COMPUTE SCORE IN COMMPASS DATA**
```{r}
DAT = exp_cbat # working from ComBat adjusted expression estimates in baseline samples

genes_17 = c("ADSS","BIRC5","CACYBP","CCT2","CCT3","CKS1B","CTPS1","ENO1",
             "GAPDH","KIAA0101","MSH6","NONO","PRKDC","RAN","SF3B4","TFB2M","UBE2A")
print(paste0("17 genes: ",length(intersect(genes_17,colnames(DAT))),
             " of ",length(genes_17)," genes in dataset"))
coeff_17 = c(0.67,0.39,0.56,1.15,0.61,0.59,0.59,0.84,0.49,
             0.30,0.91,1.03,1.04,0.65,1.01,0.83,0.66)
bjh17 = data.table(genes=genes_17,coeff=coeff_17)
# select genes in data
anno_17 = DAT %>% dplyr::select(intersect(genes_17,colnames(DAT))) 

DAT$score = data.matrix(anno_17) %*%
  diag(bjh17[genes%in%intersect(genes_17,colnames(DAT))]$coeff) %>%
  data.table() %>% rowSums()

# Check algorithm
#coe = bjh17[genes%in%intersect(genes_17,colnames(DAT))][,coeff]
#tmp = DAT[SEQ_ID=="MMRF_1024_1_BM"] %>% select(intersect(genes_17,colnames(DAT)))
#sum(tmp*coe) == DAT[SEQ_ID=="MMRF_1024_1_BM"]$score
#tmp = DAT[50,] %>% select(intersect(genes_17,colnames(DAT)))
#sum(tmp*coe) == DAT[50,]$score
```

2.2. **LINEAR REGRESSION WITH SPECTRA**
```{r}
dt.17 = data.table(merge(DAT[,c("SEQ_ID","score")],spectra))

# run linear regression with risk score as dependent variable
lm.17 = lm(data = dt.17[,-c("SEQ_ID")], formula = score ~ .) 
summary(lm.17)

nsig = data.table(summary(lm.17)$coeff[-1,"Pr(>|t|)"]) %>% 
  subset(V1<0.05) %>% nrow() # Count sig spectra in model
print(paste0(nsig," of the ",ncol(dt.17[,-c("SEQ_ID","score")]),
             " spectra were significat at p < 0.05 in the combined model"))

ajr2 = round(summary(lm.17)$adj.r.squared,digits = 2) # Print adjusted R^2
df1 = round(summary(lm.17)$fstatistic[2])
df2 = round(summary(lm.17)$fstatistic[3])
f = round(summary(lm.17)$fstatistic[1],digits = 1)
# Compute p-value from f-statistic
p = formatC(pf(f,df1,df2,lower.tail = F), format = "E", digits = 1) 
print(paste0("adjusted R^2=",ajr2,", F(",df1,",",df2,")=",f,", p=",p))

# Overall p-value
df1 = summary(lm.17)$fstatistic[2]
df2 = summary(lm.17)$fstatistic[3]
f = summary(lm.17)$fstatistic[1]
lm.17$p = pf(f,df1,df2,lower.tail = F) # Compute p-value from f-statistic
```

2.3. PLOT

Actual v predicted values
```{r}
dt = dt.17
fit = lm.17
actual_preds <- data.table(cbind(dt[,"SEQ_ID"],
                                 actual=fit$model$score,
                                 predicted=fit$fitted.values))
cor(actual_preds[,-"SEQ_ID"])
# mean absolute percentage error
mean(abs((actual_preds$predicted - actual_preds$actual))/actual_preds$actual) 
```

Plot fitted values x actual values
```{r}
p1 = ggplot(actual_preds,aes(y = predicted, x = actual)) +
  geom_point(color = "#FF000060") +
  xlab("SBUK-17 Score") + ylab("SBUK-17 PSL Score")
p1
ggsave(filename = "plots/SBUK-17_dot.pdf",plot = p1,width=3.75,height=3.75)
```

Waterfall High/Low risk score
"At the next step, the patients were dichotomized into a high‐risk and a low‐risk category based on the 75th percentile of the risk scores and survival time was compared between the two groups using the Kaplan–Meier analysis and the log‐rank test at a P value of less than 0.01. SPSS 16.0 and Survival (http://cran.r‐project.org/package=survival) packages were used to execute the survival analysis. The 75th percentile cut‐off was based on the proportion of patients in the training set who had a survival time of more than two years." (DOI: 10.1111/bjh.16744)
```{r}
# find 75th percentile of SBUK-17 score
high.cut = quantile(actual_preds$actual,.75)
# bin patients in high- or low-risk groups
actual_preds$bin = as.factor(if_else(actual_preds$actual<high.cut,"low","high"))
summary(actual_preds)

setorder(actual_preds,-predicted)
# predicted score - high risk cutoff from clustering
w <- ggplot(actual_preds,aes(y=predicted-high.cut,x=1:nrow(actual_preds),fill=bin)) + 
  geom_bar(stat="identity") +
  scale_fill_manual(values = c("red","gray"), name = "SBUK-17 Risk Group", 
                    labels = c("High (n=192)", "Low (n=575)")) +
  xlab("Patients (N=767)") + ylab("SBUK-17 PSL Score") +
  scale_x_continuous(limits = c(0,nrow(actual_preds)+1), 
                     breaks=seq(1,nrow(actual_preds),1)) +
  theme(legend.position = c(1,1), legend.justification = c(1,1),
        axis.text = element_blank(), axis.ticks = element_blank())
w
ggsave(filename = "plots/SBUK-17_waterfall.pdf",plot = w,width=3.75,height=2)
```

1.3.1 **High v low Survival Kaplan-Meier curves**
```{r}
pdt = merge(actual_preds[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttcos","censos")])
pdt$yrs = pdt$ttcos/365

fit = survfit(Surv(time = pdt$yrs, event = pdt$censos) ~ bin, data=pdt)
temp.surv = summary(fit,times=c(1,3,5))
data.table(time=temp.surv$time,strata=temp.surv$strata,survival=temp.surv$surv)

p1 = ggsurvplot(survfit(Surv(time = pdt$yrs, event = pdt$censos) ~ bin, data=pdt),
           conf.int = TRUE,
           break.x.by = 1, xlab = "years",
           ggtheme = theme_set(theme_classic() + 
                                 theme(axis.title=element_text(size=11),
                                       axis.text=element_text(size=9),
                                       panel.grid.major.x = element_line(
                                         colour = "gray",linetype = 3)
                                 )
                               )
           )
p1

survdiff(formula = Surv(time = pdt$ttcos, event = pdt$censos) ~ bin, data=pdt)
```
High/Low and TTF
```{r}
pdt = merge(actual_preds[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttctf1","censtf1")])
pdt$yrs = pdt$ttctf1/365

fit = survfit(Surv(time = pdt$yrs, event = pdt$censtf1) ~ bin, data=pdt)
temp.surv = summary(fit,times=c(1,3,5))
data.table(time=temp.surv$time,strata=temp.surv$strata,survival=temp.surv$surv)

p1 = ggsurvplot(survfit(Surv(time = pdt$yrs, event = pdt$censtf1) ~ bin, data=pdt),
           conf.int = TRUE,
           break.x.by = 1, xlab = "years",
           ggtheme = theme_set(theme_classic() + 
                                 theme(axis.title=element_text(size=11),
                                       axis.text=element_text(size=9),
                                       panel.grid.major.x = element_line(
                                         colour = "gray",linetype = 3)
                                 )
                               )
           )
p1

survdiff(formula = Surv(time = pdt$ttctf1, event = pdt$censtf1) ~ bin, data=pdt)
```

**OS at 1 year**
```{r}
dat = merge(actual_preds[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttcos","censos")])
dat$category = if_else(dat$ttcos<=365 & dat$censos==1,"dead",
                       if_else(dat$ttcos>365,"alive","censored")) 

# number of events
evts = dat %>% filter(censos==1) %>% group_by(bin) %>% 
  count(censos,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
#cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("SBUK high/low OS HR at 1 year = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```
**OS at 3 years**
```{r}
dat = merge(actual_preds[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttcos","censos")])
dat$category = if_else(dat$ttcos<=365*3 & dat$censos==1,"dead",
                       if_else(dat$ttcos>365*3,"alive","censored")) 

# number of events
evts = dat %>% filter(censos==1) %>% group_by(bin) %>% 
  count(censos,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
#cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("SBUK high/low OS HR at 3 years = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```

**TTF at 1 year**
```{r}
dat = merge(actual_preds[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttctf1","censtf1")])
dat$category = if_else(dat$ttctf1<=365 & dat$censtf1==1,"dead",
                       if_else(dat$ttctf1>365,"alive","censored")) 

# number of events
evts = dat %>% filter(censtf1==1) %>% group_by(bin) %>% 
  count(censtf1,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
#cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("SBUK high/low TTF HR at 1 year = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```

**TTF at 3 years**
```{r}
dat = merge(actual_preds[,c("SEQ_ID","bin")],dt.all[,c("SEQ_ID","ttctf1","censtf1")])
dat$category = if_else(dat$ttctf1<=365*3 & dat$censtf1==1,"dead",
                       if_else(dat$ttctf1>365*3,"alive","censored")) 

# number of events
evts = dat %>% filter(censtf1==1) %>% group_by(bin) %>% 
  count(censtf1,name="events")  

# count number in each category
dead = dat %>% filter(category=="dead") %>% group_by(bin) %>% 
  count(category,name = "n.dead")
aliv = dat %>% filter(category=="alive") %>% group_by(bin) %>% 
  count(category,name = "n.alive")
cens = dat %>% filter(category=="censored") %>% group_by(bin) %>% 
  count(category,name="n.censored")
cnt = data.table(
  merge(evts,
        merge(dead,
              merge(aliv,cens,by="bin",all = T)
              ,by="bin",all=T)
        ,by="bin",all=T)
)[,c("bin","events","n.dead","n.alive","n.censored")]
cnt


a = cnt[bin%in%"high",n.dead]
b = cnt[bin%in%"high",n.alive]
c = sum(cnt[bin%in%"low",n.dead])
d = sum(cnt[bin%in%"low",n.alive])

hr = (a/b)/(c/d)
se = sqrt((1/a)+(1/b)+(1/c)+(1/d))
cil = exp(log(hr)-1.96*se)
ciu = exp(log(hr)+1.96*se)

paste0("SBUK high/low TTF HR at 3 years = ",
       round(hr,digits = 2)," (",
       round(cil,digits = 2),"-",
       round(ciu,digits = 2),")")

```

**Save model results**
```{r}
save(lm.17,lm.70,file = "rdata/lm.expression-scores.rdata")
```
