==> 3.cd138_spectra.clinical_risk.Rmd <==
---
title: "CD138+ spectra and clinical risk"
author: "Rosalie Griffin Waller"
date: "27-OCT-2020"
output:
  pdf_document: default
  html_document:
    theme: united
    highlight: tango
---
Code to investigate associations between CD138+ spectra and clinical risk cytogenetics and disease stage

### 0. Setup
```{r load_packages, message=F, include=T,error=F,warning=F}
# Install and load required R packages
library(ggplot2)
library(MASS)
library(survivalAnalysis)
library(dplyr)
library(data.table)
```

Define data directory
```{r}
data_dir = "/path/to/data" # exclude ending "/"
```
```{r, include=FALSE}
# real data directory
data_dir = "/Users/rosal/OneDrive - University of Utah/2020/analyze/data/spectra.v2/data"
```

Load transcriptome spectra (PC1-PC)
```{r}
# read in spectra and clinical data
spectra_clinical = read.csv(file = paste0(data_dir,
                                  "/baseline-clinical-spectra-sd.csv")
                            ) %>% data.table()
```

### 1. Cytogenetics
Large somatic chromosomal DNA aberrations detected by cytogenetics are used to define prognostic risk groups in myeloma. Clinical risk categories defined by mSMART7 include: high risk, del(17p) and t(14;16); intermediate risk, amp(1q) and t(4;14); and standard risk, t(11;14). 

1.1. del(17p): D_TRI_CF_ABNORMALITYPR11 & D_TRI_CF_17PABNORMALCE (abundance)
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_TRI_CF_ABNORMALITYPR11",starts_with("PC")) %>%
  dplyr::filter(D_TRI_CF_ABNORMALITYPR11!="Not Done" &
                D_TRI_CF_ABNORMALITYPR11!="")

print(paste("Baseline samples with data:",nrow(DAT),"(total),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR11=="Yes",]),"(+),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR11=="No",]),"(-)"))
DAT$D_TRI_CF_ABNORMALITYPR11 = as.factor(DAT$D_TRI_CF_ABNORMALITYPR11)

# Logistic regression
MOD = glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR11 ~ .,family = "binomial")
summary(MOD)

# Count number of significant spectra
nsig = data.table(summary(MOD)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Save model
mod.risk = list(d17p=MOD)

# Overall p-value
NLL = glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR11 ~ 1,family = "binomial")

mod.risk$d17p$p = pchisq(deviance(NLL)-deviance(MOD),
                         df.residual(NLL)-df.residual(MOD),
                         lower.tail=FALSE)
mod.risk$d17p$p
```

Run model with significant spectra only
```{r}
md2 = glm(data = DAT, 
          formula = D_TRI_CF_ABNORMALITYPR11 ~ PC3_SD + PC6_SD + PC39_SD, 
          family = "binomial")
summary(md2)
pchisq(deviance(NLL)-deviance(md2),
       df.residual(NLL)-df.residual(md2),lower.tail=FALSE)
```

1.2. t(14;16): D_TRI_CF_ABNORMALITYPR8 & D_TRI_CF_T1416ABNORMAL (abundance)
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_TRI_CF_ABNORMALITYPR8",starts_with("PC")) %>%
  dplyr::filter(D_TRI_CF_ABNORMALITYPR8!="Not Done" &
                D_TRI_CF_ABNORMALITYPR8!="")

print(paste("Baseline samples with data:",nrow(DAT),"(total),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR8=="Yes",]),"(+),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR8=="No",]),"(-)"))

DAT$D_TRI_CF_ABNORMALITYPR8 = as.factor(DAT$D_TRI_CF_ABNORMALITYPR8)

# Logistic regression
MOD <- glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR8 ~ .,family = "binomial")
summary(MOD)

# Count sig spectra in model
nsig = data.table(summary(MOD)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Save model
mod.risk = c(mod.risk,list(t1416=MOD))

# Overall p-value
NLL = glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR8 ~ 1,family = "binomial")

mod.risk$t1416$p = pchisq(deviance(NLL)-deviance(MOD),
                         df.residual(NLL)-df.residual(MOD),
                         lower.tail=FALSE)
mod.risk$t1416$p
```


1.3. amp(1q): D_TRI_CF_ABNORMALITYPR13 & D_TRI_CF_1PAMPLIFICATI2 (abundance)
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_TRI_CF_ABNORMALITYPR13",starts_with("PC")) %>%
  dplyr::filter(D_TRI_CF_ABNORMALITYPR13!="Not Done" &
                D_TRI_CF_ABNORMALITYPR13!="")

print(paste("Baseline samples with data:",nrow(DAT),"(total),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR13=="Yes",]),"(+),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR13=="No",]),"(-)"))
DAT$D_TRI_CF_ABNORMALITYPR13 = as.factor(DAT$D_TRI_CF_ABNORMALITYPR13)

# Logistic regression
MOD <- glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR13 ~ .,family = "binomial")
summary(MOD)

# Count sig spectra in model
nsig = data.table(summary(MOD)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Save model
mod.risk = c(mod.risk,list(a1q=MOD))

# Overall p-value
NLL = glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR13 ~ 1,family = "binomial")

mod.risk$a1q$p = pchisq(deviance(NLL)-deviance(MOD),
                         df.residual(NLL)-df.residual(MOD),
                         lower.tail=FALSE)
mod.risk$a1q$p
```


1.4. t(4;14): D_TRI_CF_ABNORMALITYPR3 & D_TRI_CF_T414ABNORMALC (abundance)
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_TRI_CF_ABNORMALITYPR3",starts_with("PC")) %>%
  dplyr::filter(D_TRI_CF_ABNORMALITYPR3!="Not Done" &
                D_TRI_CF_ABNORMALITYPR3!="")

print(paste("Baseline samples with data:",nrow(DAT),"(total),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR3=="Yes",]),"(+),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR3=="No",]),"(-)"))
DAT$D_TRI_CF_ABNORMALITYPR3 = as.factor(DAT$D_TRI_CF_ABNORMALITYPR3)

# Logistic regression
MOD <- glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR3 ~ .,family = "binomial")
summary(MOD)

# Count sig spectra in model
nsig = data.table(summary(MOD)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Save model
mod.risk = c(mod.risk,list(t414=MOD))

# Overall p-value
NLL = glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR3 ~ 1,family = "binomial")

mod.risk$t414$p = pchisq(deviance(NLL)-deviance(MOD),
                         df.residual(NLL)-df.residual(MOD),
                         lower.tail=FALSE)
mod.risk$t414$p
```

1.5. t(11;14): D_TRI_CF_ABNORMALITYPR6 & D_TRI_CF_T1114ABNORMAL (abundance)
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_TRI_CF_ABNORMALITYPR6",starts_with("PC")) %>%
  dplyr::filter(D_TRI_CF_ABNORMALITYPR6!="Not Done" &
                D_TRI_CF_ABNORMALITYPR6!="")

print(paste("Baseline samples with data:",nrow(DAT),"(total),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR6=="Yes",]),"(+),",
            nrow(DAT[DAT$D_TRI_CF_ABNORMALITYPR6=="No",]),"(-)"))
DAT$D_TRI_CF_ABNORMALITYPR6 = as.factor(DAT$D_TRI_CF_ABNORMALITYPR6)

# Logistic regression
MOD <- glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR6 ~ .,family = "binomial")
summary(MOD)

# Count sig spectra in model
nsig = data.table(summary(MOD)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Save model
mod.risk = c(mod.risk,list(t1114=MOD))

# Overall p-value
NLL = glm(data = DAT,formula = D_TRI_CF_ABNORMALITYPR6 ~ 1,family = "binomial")

mod.risk$t1114$p = pchisq(deviance(NLL)-deviance(MOD),
                         df.residual(NLL)-df.residual(MOD),
                         lower.tail=FALSE)
mod.risk$t1114$p
```


### 2. International staging system (ISS)
2.1. RUN LOGISTIC REGRESSION
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_PT_iss",starts_with("PC"))

# Order dependent variable
DAT$ISS = factor(DAT$D_PT_iss, levels = c(1,2,3), ordered = TRUE)

# Ordinal logistic regression model
MOD = polr(data = DAT[,-c("D_PT_iss")], formula = ISS ~ ., Hess = T) 
summary(MOD)

# Count sig spectra in model
sig = pnorm(abs(coef(summary(MOD))[1:39,"t value"]),lower.tail = FALSE)*2
nsig = data.table(sig) %>% subset(sig<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Save model
mod.risk = c(mod.risk,list(iss=MOD))

# Overall p-value
NLL = polr(data = DAT[,-c("D_PT_iss")], formula = ISS ~ 1, Hess = T)

mod.risk$iss$p = pchisq(deviance(NLL)-deviance(MOD),
                         df.residual(NLL)-df.residual(MOD),
                         lower.tail=FALSE)
mod.risk$iss$p
```
**Save model results**
```{r}
save(mod.risk,file = "rdata/mod.clinical-risk.rdata")
```
