==> 4.cd138_spectra.disease_course.Rmd <==
---
title: "CD138+ spectra and disease course"
output:
  pdf_document: default
  html_document:
    theme: united
    highlight: tango
---

**Test for associations between spectrum and overall survival (OS) and time to first-line treatment failure (TTF) in the MMRF CoMMpass data.**


#### 0. Setup

Define data directory
```{r}
data_dir = "/path/to/data" # exclude ending "/"
```
```{r, include=FALSE}
# real data directory
data_dir = "/Users/rosal/OneDrive - University of Utah/2020/analyze/data/spectra.v2/data"
```

Load packages
```{r load_packages, message=F, include=T,error=F,warning=F}
# Install and load required R packages
library(dplyr)
library(data.table)
library(ggplot2)
library(MASS)
library(survivalAnalysis)
```

Load spectra (PC1-PC39) and clinical data
```{r}
dt.all = read.csv(paste0(data_dir,"/baseline-clinical-spectra-sd.csv"))
```

#### 1. Overall Surival (OS)
```{r,warning=F}
dt.os = dt.all %>% dplyr::select("SEQ_ID","ttcos","censos",starts_with("PC"))
dt.os$censos = as.integer(dt.os$censos) # set variable type
os = analyse_multivariate(data = dt.os,
                           time_status = vars(ttcos,censos),
                           covariates = colnames(dplyr::select(dt.os,
                                                               ends_with("_SD"))))
summary(os$coxph)
summary(os$coxph)$logtest

# count spectrum in model with p < .05
nsig = data.table(summary(os$coxph)$coeff[,"Pr(>|z|)"]) %>% 
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," spectra significant at p<.05"))
```

#### 2. Time to first-line treatment failure (TTF)
```{r}
dt.tf = dt.all %>% dplyr::select("SEQ_ID","ttctf1","censtf1",starts_with("PC"))
dt.tf$censtf1 = as.integer(dt.tf$censtf1) # set variable type
tf <- analyse_multivariate(data = dt.tf,
                           time_status = vars(ttctf1,censtf1),
                           covariates = colnames(dplyr::select(dt.tf,
                                                               ends_with("_SD"))))
summary(tf$coxph)
summary(tf$coxph)$logtest

# count spectrum in model with p < .05
nsig = data.table(summary(tf$coxph)$coeff[,"Pr(>|z|)"]) %>% 
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," spectra significant at p<.05"))
```

**Save model results**
```{r}
save(tf,os,file = "rdata/mod.disease-course.rdata")
```
