==> 5.cd138_spectra.demographic_risk.Rmd <==
---
title: "CD138+ spectra and demographic risk groups"
output:
  pdf_document: default
  html_document:
    theme: united
    highlight: tango
---
**Code to investigate associations between patient demographics (age, gender, self-reported race and ethnicity) and transcriptome spectra using analysis of variance.**

### 0. Setup

Define data directory
```{r}
data_dir = "/path/to/data" # exclude ending "/"
```
```{r, include=FALSE}
# real data directory
data_dir = "/Users/rosal/OneDrive - University of Utah/2020/analyze/data/spectra.v2/data"
```

Load packages
```{r load_packages, message=F, include=T,error=F,warning=F}
# Install and load required R packages
library(dplyr)
library(data.table)
library(ggplot2)
library(MASS)
library(survivalAnalysis)
```

Load transcriptome spectra (PC1-PC39)
```{r}
# read in spectra and clinical data
spectra_clinical = read.csv(file = paste0(data_dir,
                                  "/baseline-clinical-spectra-sd.csv")
                            ) %>% data.table()
```

### 1. Age
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_PT_age",starts_with("PC"))

# Linear regression
lm.age = lm(data = DAT,formula = D_PT_age ~ .)
summary(lm.age)

# Count number of significant spectra
nsig = data.table(summary(lm.age)$coeff[-1,"Pr(>|t|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Overall p-value
df1 = summary(lm.age)$fstatistic[2]
df2 = summary(lm.age)$fstatistic[3]
f = summary(lm.age)$fstatistic[1]
lm.age$p = pf(f,df1,df2,lower.tail = F) # Compute p-value from f-statistic
```

### 2. Gender
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_PT_gender",starts_with("PC"))
DAT$D_PT_gender = as.factor(DAT$D_PT_gender)

# Logistic regression
glm.gender = glm(data = DAT,formula = D_PT_gender ~ .,family = "binomial")
summary(glm.gender)

# Count number of significant spectra
nsig = data.table(summary(glm.gender)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Overall p-value
NLL = glm(data = DAT,formula = D_PT_gender ~ 1,family = "binomial")

glm.gender$p = pchisq(deviance(NLL)-deviance(glm.gender),
                         df.residual(NLL)-df.residual(glm.gender),
                         lower.tail=FALSE)
```

### 3. Self-reported race
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_PT_race",starts_with("PC")) %>%
  dplyr::filter(D_PT_race%in%c(1,2))
DAT$D_PT_race = as.factor(DAT$D_PT_race)

# Logistic regression
glm.race = glm(data = DAT,formula = D_PT_race ~ .,family = "binomial")
summary(glm.race)

# Count number of significant spectra
nsig = data.table(summary(glm.race)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Overall p-value
NLL = glm(data = DAT,formula = D_PT_race ~ 1,family = "binomial")

glm.race$p = pchisq(deviance(NLL)-deviance(glm.race),
                         df.residual(NLL)-df.residual(glm.race),
                         lower.tail=FALSE)
```

### 4. Self-reported ethnicity
```{r}
DAT = spectra_clinical %>% 
  dplyr::select("D_PT_ethnic",starts_with("PC")) %>%
  dplyr::filter(D_PT_ethnic%in%c(1,2))
DAT$D_PT_ethnic = as.factor(DAT$D_PT_ethnic)

# Logistic regression
glm.ethnic = glm(data = DAT,formula = D_PT_ethnic ~ .,family = "binomial")
summary(glm.ethnic)

# Count number of significant spectra
nsig = data.table(summary(glm.ethnic)$coeff[-1,"Pr(>|z|)"]) %>%
  subset(V1<0.05) %>% nrow()
print(paste0(nsig," of 39 spectra significant (p<.05)"))

# Overall p-value
NLL = glm(data = DAT,formula = D_PT_ethnic ~ 1,family = "binomial")

glm.ethnic$p = pchisq(deviance(NLL)-deviance(glm.ethnic),
                         df.residual(NLL)-df.residual(glm.ethnic),
                         lower.tail=FALSE)
```
**Save model results**
```{r}
save(lm.age,glm.gender,glm.race,glm.ethnic,file = "rdata/mod.demographic-risk.rdata")
```
